+++ Source/dataProcessing/mathGenerators.h
@@ -1,0 +1,85 @@
+#include <JuceHeader.h>
+enum FFTOrder
+{
+    order2048 = 11,
+    order4096 = 12,
+    order8192 = 13
+};
+
+template<typename BlockType>
+struct FFTDataGenerator
+{
+    /**
+     produces the FFT data from an audio buffer.
+     */
+    void produceFFTDataForRendering(const juce::AudioBuffer<float>& audioData, const float negativeInfinity)
+    {
+        const auto fftSize = getFFTSize();
+
+        fftData.assign(fftData.size(), 0);
+        auto* readIndex = audioData.getReadPointer(0);
+        std::copy(readIndex, readIndex + fftSize, fftData.begin());
+
+        // first apply a windowing function to our data
+        window->multiplyWithWindowingTable(fftData.data(), fftSize);       // [1]
+
+        // then render our FFT data..
+        forwardFFT->performFrequencyOnlyForwardTransform(fftData.data());  // [2]
+
+        int numBins = (int)fftSize / 2;
+
+        //normalize the fft values.
+        for (int i = 0; i < numBins; ++i)
+        {
+            auto v = fftData[i];
+            //            fftData[i] /= (float) numBins;
+            if (!std::isinf(v) && !std::isnan(v))
+            {
+                v /= float(numBins);
+            }
+            else
+            {
+                v = 0.f;
+            }
+            fftData[i] = v;
+        }
+
+        //convert them to decibels
+        for (int i = 0; i < numBins; ++i)
+        {
+            fftData[i] = juce::Decibels::gainToDecibels(fftData[i], negativeInfinity);
+        }
+
+        fftDataFifo.push(fftData);
+    }
+
+    void changeOrder(FFTOrder newOrder)
+    {
+        //when you change order, recreate the window, forwardFFT, fifo, fftData
+        //also reset the fifoIndex
+        //things that need recreating should be created on the heap via std::make_unique<>
+
+        order = newOrder;
+        auto fftSize = getFFTSize();
+
+        forwardFFT = std::make_unique<juce::dsp::FFT>(order);
+        window = std::make_unique<juce::dsp::WindowingFunction<float>>(fftSize, juce::dsp::WindowingFunction<float>::blackmanHarris);
+
+        fftData.clear();
+        fftData.resize(fftSize * 2, 0);
+
+        fftDataFifo.prepare(fftData.size());
+    }
+    //==============================================================================
+    int getFFTSize() const { return 1 << order; }
+    int getNumAvailableFFTDataBlocks() const { return fftDataFifo.getNumAvailableForReading(); }
+    //==============================================================================
+    bool getFFTData(BlockType& fftData) { return fftDataFifo.pull(fftData); }
+private:
+    FFTOrder order;
+    BlockType fftData;
+    std::unique_ptr<juce::dsp::FFT> forwardFFT;
+    std::unique_ptr<juce::dsp::WindowingFunction<float>> window;
+
+    Fifo<BlockType> fftDataFifo;
+};
